requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langchain-openai
langchain-pinecone
langgraph
ai-sdk-python
python-dotenv
ollama

